[RL] Fix the incorrect routing of EOS tokens, which leads to changes in accuracy (#7960)
* Reset buffer size of R3

* refine code

* R3 fix Eos bug

* pre-commit

* fix r3 ci and support dsa

* refine code

* refine code

* reset ci dir

* refine code

* fix dsv3

+++ fastdeploy/model_executor/models/deepseek_v3.py
@@ -1085,7 +1085,7 @@ class DeepSeekV3Model(nn.Layer):
                 forward_meta,
                 hidden_states,
                 residual,
-                position_ids,
+                position_ids.cast(paddle.int32),
             )
         out = self.norm(hidden_states, residual, forward_meta=forward_meta)[0]
